remove pge/product references from `s1_disp`
also move 2 functiosn for `disp-s1` into `_utils`
